tools_QGIS/dbQGIS/databricks_keepalive2_requirements.txt
# Requirements for databricks_keepalive2
python-dotenv
databricks-sql-connector
